requirements.txt
akshare>=1.12.0
pandas>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
